+++ GrowthBook/src/commonMain/kotlin/com/sdk/growthbook/evaluators/GBFeatureEvaluator.kt
@@ -168,13 +168,12 @@ internal class GBFeatureEvaluator(
                         /**
                          * If it's a conditional rule, skip if the condition doesn't pass
                          */
-                        if (rule.condition != null && !GBConditionEvaluator().evalCondition(
+                        if (rule.conditionGB != null && !GBConditionEvaluator().evalCondition(
                                 attributes = getAttributes(
                                     attributeOverrides = attributeOverrides,
                                     attributes = evaluationContext.userContext.attributes,
                                 ),
-                                conditionObj = rule.condition.let(GBValue::from) as? GBJson
-                                    ?: GBJson(emptyMap()),
+                                conditionObj = rule.conditionGB,
                                 savedGroups = evaluationContext.savedGroups,
                             )
                         ) {
@@ -309,7 +308,8 @@ internal class GBFeatureEvaluator(
                                 .evaluateExperiment(
                                     featureId = featureKey,
                                     experiment = exp,
-                                    attributeOverrides = attributeOverrides
+                                    attributeOverrides = attributeOverrides,
+                                    conditionObj = rule.conditionGB,
                                 )
                             if (result.inExperiment && (result.passthrough != true)) {
                                 return prepareResult(